% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{aggregate_band_powers}
\alias{aggregate_band_powers}
\title{aggregate_band_powers}
\usage{
aggregate_band_powers(hypnogram_band_powers)
}
\arguments{
\item{hypnogram_band_powers}{hypnogram_band_powers.}
}
\value{
Wide dataframe.
}
\description{
aggregate_band_powers
}
