% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ah_back}
\alias{ah_back}
\title{Apnea and Hypopnea on back}
\usage{
ah_back(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Apnea and Hypopnea on back
}
