% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ah_count}
\alias{ah_count}
\title{Apnea and Hypopnea count}
\usage{
ah_count(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Apnea and Hypopnea count
}
