% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ah_hour}
\alias{ah_hour}
\title{Apnea and Hypopnea index}
\usage{
ah_hour(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Apnea and Hypopnea index
}
