% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ah_nonback}
\alias{ah_nonback}
\title{Apnea and Hypopnea non back}
\usage{
ah_nonback(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Apnea and Hypopnea non back
}
