% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ah_rem}
\alias{ah_rem}
\title{Apnea and Hypopnea in rem}
\usage{
ah_rem(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Apnea and Hypopnea in rem
}
