% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{awa_minutes}
\alias{awa_minutes}
\title{Get total duration of AWA in minutes.}
\usage{
awa_minutes(hypnogram)
}
\arguments{
\item{hypnogram}{Hypnogram dataframe.}
}
\value{
total duration of N3 sleep in minutes.
}
\description{
Get total duration of AWA in minutes.
}
