% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{bands.power}
\alias{bands.power}
\title{Campute bands powers.}
\usage{
bands.power(x, Fs, bands = list(c(0.5, 3.5), c(3.5, 8), c(8, 12), c(12,
  30), c(30, 50)), bands.names = c("delta", "theta", "alpha", "beta",
  "gamma1"))
}
\arguments{
\item{x}{signal vector.}

\item{Fs}{Frequency.}

\item{bands}{bands. Default "delta","theta","alpha","beta", "gamma1".}
}
\value{
a named vector with all bands powers.
}
\description{
Campute bands powers.
}
