% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borbely.R
\name{borbely}
\alias{borbely}
\title{Solve Borbely's model}
\usage{
borbely(ts, y0, awake0 = FALSE, parms = borbely_default_parms())
}
\arguments{
\item{ts}{Vector of times (in h)}

\item{y0}{Initial condition}

\item{awake0}{Initial state (TRUE for awake, FALSE for asleep)}

\item{parms}{Model parameters (optional, see \code{\link{borbely_default_parms}})}
}
\value{
Results of the simulation, including times, states and asleep/awake status
}
\description{
Solves the Borbely's sleep-wake model for the given times, initial condition and parameters
}
\examples{
ts <- seq(0, 10, length.out = 2000)
sol <- borbely(seq(0, 10, length.out = 2000), 1)
}
\references{
Skeldon AC, Dijk D-J, Derks G. Mathematical Models for Sleep-Wake Dynamics:
Comparison of the Two-Process Model and a Mutual Inhibition Neuronal Model.
PLoS One. 2014 Aug 1;9(8):e103877. Available from: http://dx.plos.org/10.1371/journal.pone.0103877
}
\seealso{
\code{\link{borbely_default_parms}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
