% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borbely.R
\name{borbely_default_parms}
\alias{borbely_default_parms}
\title{Default parameters of Borbely's model}
\usage{
borbely_default_parms()
}
\value{
The default parameters for Borbely's model
}
\description{
Loads the parameters used in Borbely's model
}
\examples{
parms <- borbely_default_parms()
}
\references{
Skeldon AC, Dijk D-J, Derks G. Mathematical Models for Sleep-Wake Dynamics:
Comparison of the Two-Process Model and a Mutual Inhibition Neuronal Model.
PLoS One. 2014 Aug 1;9(8):e103877. Available from: http://dx.plos.org/10.1371/journal.pone.0103877
}
\seealso{
\code{\link{borbely}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
