% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borbely.R
\name{borbelyPlot}
\alias{borbelyPlot}
\title{Plots the time series generated with borbely}
\usage{
borbelyPlot(sol)
}
\arguments{
\item{sol}{Solution of a borbely simulation}
}
\value{
The simulation's plot
}
\description{
Plots the time series generated with borbely
}
\examples{
# Generate a solution
y0 <- 0.4
nDays <- 3
ts <- seq(0, nDays*24, length.out = nDays*24*20)
sol <- borbely(ts, y0)
# Plot it
borbelyPlot(sol)
}
\references{
Skeldon AC, Dijk D-J, Derks G. Mathematical Models for Sleep-Wake Dynamics:
Comparison of the Two-Process Model and a Mutual Inhibition Neuronal Model.
PLoS One. 2014 Aug 1;9(8):e103877. Available from: http://dx.plos.org/10.1371/journal.pone.0103877
}
\seealso{
\code{\link{borbely}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
