% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout-analysis.R
\name{bout_analysis_standard}
\alias{bout_analysis_standard}
\title{Standard evaluation of bout_analysis}
\usage{
bout_analysis_standard(var_name, data)
}
\arguments{
\item{var_name}{character, name of the column in data to be processed}

\item{data}{\link{data.table} containing behavioural variable from or one multiple animals.
When it has a key, unique values, are assumed to represent unique individuals (e.g. in a \link{behavr} table).
Otherwise, it analysis the data as coming from a single animal. \code{data} must have a column \code{t} representing time.}
}
\description{
Standard evaluation of bout_analysis
}
\seealso{
bout_analysis
}
