% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_all_stats}
\alias{compute_all_stats}
\title{Call all available stats functions}
\usage{
compute_all_stats(records, eeg_channels = c("C3-A2", "EEG Fpz-Cz"))
}
\arguments{
\item{eeg_channels}{potential EEG channel names.}

\item{record}{record paths.}
}
\value{
df.
}
\description{
Call all available stats functions
}
