% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_awards.R
\name{create_awards}
\alias{create_awards}
\title{Create Awards}
\usage{
create_awards(league_id = "1124848060283768832", nfl_start_dt = "2024-09-05")
}
\arguments{
\item{league_id}{A string representing the ID of the fantasy football league (default: "1124848060283768832").}

\item{nfl_start_dt}{A string representing the start date of the NFL season (default: "2024-09-05").}
}
\value{
A dataframe (tibble) containing awards, with the following columns:
\describe{
\item{\code{award_name}}{The name of the award.}
\item{\code{award_winner}}{The name of the award recipient.}
\item{\code{award_description}}{A brief description of the award.}
\item{\code{award_details}}{Detailed information about why the award was given.}
}
}
\description{
Generates a comprehensive set of awards for a fantasy football league based on team and player performance.
}
\details{
This function analyzes league data to calculate and assign awards for achievements like the highest individual
scoring player, most efficient manager, smallest margin of loss, and many others. The function processes weekly points,
aggregated team data, and player metadata to identify award winners and provide detailed descriptions of their accomplishments.
}
\examples{
\dontrun{
# Generate awards for a league
awards_df <- create_awards(league_id = "1124848060283768832", nfl_start_dt = "2024-09-05")
print(awards_df)
}
}
