% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_long_points_df.R
\name{create_long_points_df}
\alias{create_long_points_df}
\title{Create Long Points Dataframe}
\usage{
create_long_points_df(
  league_id = "1124848060283768832",
  nfl_start_dt = "2024-09-05",
  player_df = NA
)
}
\arguments{
\item{league_id}{A string representing the ID of the fantasy football league (default: "1124848060283768832").}

\item{nfl_start_dt}{A string representing the start date of the NFL season (default: "2024-09-05").}

\item{player_df}{A dataframe containing player metadata. If not provided, the function dynamically retrieves player data using \code{get_player_df}.}
}
\value{
A dataframe (tibble) containing long-format player points data, including:
\describe{
\item{\code{week}}{The week number of the NFL season.}
\item{\code{roster_id}}{The ID of the roster.}
\item{\code{display_name}}{The display name of the user.}
\item{\code{team_name}}{The name of the team.}
\item{\code{player_id}}{The ID of the player.}
\item{\code{player_name}}{The full name of the player.}
\item{\code{player_position}}{The position of the player (e.g., QB, RB, WR, etc.).}
\item{\code{player_points}}{The points scored by the player during the week.}
\item{\code{is_starter}}{A designation indicating if the player is a starter or on the bench.}
}
}
\description{
Generates a detailed dataframe of individual player points for each roster and week in a fantasy football league.
}
\details{
This function combines weekly points, user information, roster details, and player metadata to produce a comprehensive
long-format dataframe with player-level details for each week.
}
\examples{
\dontrun{
# Fetch long-format player points data for a league
long_points_df <- create_long_points_df(
  league_id = "1124848060283768832",
  nfl_start_dt = "2024-09-05"
)
print(long_points_df)
}
}
