% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_roster_df.R
\name{create_roster_df}
\alias{create_roster_df}
\title{Create Roster Dataframe}
\usage{
create_roster_df(league_id = "1124848060283768832")
}
\arguments{
\item{league_id}{A string representing the ID of the fantasy football league (default: "1124848060283768832").}
}
\value{
A dataframe (tibble) containing roster information, including:
\describe{
\item{\code{league_id}}{The ID of the league.}
\item{\code{owner_id}}{The ID of the owner.}
\item{\code{roster_id}}{The ID of the roster.}
}
}
\description{
Fetches roster information for a specified fantasy football league and organizes it into a dataframe.
}
\details{
This function retrieves roster details from the Sleeper API for a given league ID. It creates a mapping between
\code{owner_id} and \code{roster_id}, along with the associated league ID.
}
\examples{
\dontrun{
# Fetch roster data for a league
roster_df <- create_roster_df(league_id = "1124848060283768832")
print(roster_df)
}
}
