% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_season_weekly_df.R
\name{create_season_weekly_df}
\alias{create_season_weekly_df}
\title{Create Season Weekly Dataframe}
\usage{
create_season_weekly_df(nfl_start_dt = "2024-09-05")
}
\arguments{
\item{nfl_start_dt}{A string representing the start date of the NFL season (default: "2024-09-05").
This date is used to calculate the weekly start and end dates.}
}
\value{
A list containing:
\describe{
\item{season_df}{A dataframe (tibble) with columns:
\itemize{
\item{\code{week}: The week number (1 to 18).}
\item{\code{start_dt}: The start date of each week (Thursday).}
\item{\code{end_dt}: The end date of each week (Monday).}
\item{\code{month}: The month of the start date.}
}
}
\item{current_week}{A dataframe (tibble) containing details of the current week based on today's date.}
}
}
\description{
Generates a dataframe representing weekly start and end dates for an NFL season.
}
\details{
This function creates a dataframe with details for each of the 18 NFL weeks, including
the week number, start and end dates, and the month corresponding to each week. It also
identifies the current week based on the current date.
}
\examples{
\dontrun{
# Create the season weekly dataframe and identify the current week
season_data <- create_season_weekly_df("2024-09-05")
season_df <- season_data$season_df
current_week <- season_data$current_week
print(season_df)
print(current_week)
}
}
