% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_starter_df.R
\name{create_starter_df}
\alias{create_starter_df}
\title{Create a Starter Dataframe}
\usage{
create_starter_df(roster_id, starters)
}
\arguments{
\item{roster_id}{A vector representing the ID(s) of the roster(s). This will be coerced to a character type.}

\item{starters}{A vector representing the player_ids for the starters.}
}
\value{
A tibble (dataframe) with the following columns:
\describe{
\item{\code{roster_id}}{The ID of the roster, stored as a character string.}
\item{\code{starters}}{The identifiers for the starters.}
\item{\code{is_starter}}{A constant value "starter_points" indicating the player is a starter.}
}
}
\description{
Generates a dataframe representing starters for a given roster.
}
\details{
This function creates a dataframe with information about the starters for a specific roster.
Each row represents a starter and includes the roster ID, the starter's identifier, and a
column indicating that the player is a starter.
}
\examples{
# Create a starter dataframe for a roster
create_starter_df(
  roster_id = 1,
  starters = c("11563" "9509"  "8138"  "11631" "4039"  "4217"  "7496"  "11635" "1945"  "BAL")
)

}
