% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_annotation.R
\name{movement_detector_enclosed}
\alias{movement_detector_enclosed}
\alias{custom_annotation_wrapper}
\title{Generic function to aggregate movement with some statistic}
\usage{
movement_detector_enclosed(func, feature, statistic, score, preproc_FUN = NULL)

custom_annotation_wrapper(custom_function)
}
\arguments{
\item{func}{Aggregating function (max, min, median, mean, etc)}

\item{feature}{Name of a column in the sqlite3 file e.g. xy_dist_log10x1000}

\item{statistic}{Name of the column resulting from aggregation e.g. max_movement}

\item{score}{Name of the column providing a score i.e. category to the statistic e.g. micromovement
score is usually a binary variable i.e. TRUE/FALSE}

\item{preproc_FUN}{Optional, function to preprocess the input before computing the feature
(if the data needs some transformation like reverting xy_dist_log10x1000 back to a distance)}

\item{custom_function}{function used to produce the custom annotation}

\item{data}{\link{data.table} containing behavioural variable from or one multiple animals.
When it has a key, unique values, are assumed to represent unique individuals (e.g. in a \link{behavr} table).
Otherwise, it analysis the data as coming from a single animal. \code{data} must have a column \code{t} representing time.}

\item{time_window_length}{Size of non overlapping time bins, in seconds}

\item{...}{Extra arguments to be passed to \code{custom_function}.}
}
\value{
a \link{behavr} table similar to \code{data} with additional variables/annotations.
The resulting data will only have one data point every \code{time_window_length} seconds.
}
\description{
This function gives aggregates a variable of interest in a custom way
All datapoints in every time_window_length seconds is aggregated into a single datapoint
}
\details{
movement_detector_enclosed takes:
\itemize{
\item{the name of an R summary function (mean, max, etc)}
\item{the name of a column in the future datasets to apply the function to}
\item{the name of the resulting summary column}
\item{the name of an alternative boolean column, which is set to TRUE if the summary column has a value greater than a threshold (default 1)}
\item{a preprocessing function to be applied to the column before the summary function is applied to it}
}

The default \code{time_window_length} is 300 seconds -- it is also known as the "5-minute rule".
custom_annotation_wrapper simplifies writing new annotation functions by leaving the shared functionality here
and the dedicated functionality to the new function.
This function adds to the functionality in the annotation function:
\itemize{
\item{Check a minimal amount of data is available and quit otherwise}
\item{Restore the name of the time column to remove the effects of binning}
\item{Check the amount of data after annotation is also enough (at least 1)}
\item{Apply a rolling interpolation of the labels to the data (assume the last available data point)}
}
It implements 3 attributes:
\itemize{
\item{needed_columns: A function that returns the columns needed by the function in its passed data}
\item{parameters: A function that returns the name of the parameters used by the function (including other functions' called by it)}
\item{variables: A function that returns the name of the newly produced columns by the function}
}
}
