% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cycles_end_avg_duration}
\alias{cycles_end_avg_duration}
\title{cycles_end_avg_duration}
\usage{
cycles_end_avg_duration(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
cycles_end_avg_duration
}
