% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cycles_stats}
\alias{cycles_stats}
\title{Get cycles related stats}
\usage{
cycles_stats(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Get cycles related stats
}
