% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronauer.R
\name{dKronauer}
\alias{dKronauer}
\title{Kronauer's 1990 model flow}
\usage{
dKronauer(time, y, I, parms = kronauer_default_parms())
}
\arguments{
\item{time}{The time (in h)}

\item{y}{The state}

\item{I}{Illumination function (in lux)}

\item{parms}{Model's parameters (optional, see \code{\link{kronauer_default_parms}})}
}
\value{
The flow (right hand side of the differential equation)
}
\description{
Generates the flow (right hand side of the differential equation) for the Kronauer's 1990 model (equations 2 and 3 in the reference)
}
\examples{
\dontrun{
t <- 0
y <- c(x = 1, xc = 1)
dy <- dKronauer(t, y)
}
}
\references{
Jewett ME, Kronauer RE.
Refinement of Limit Cycle Oscillator Model of the Effects of Light on the Human Circadian Pacemaker.
J Theor Biol. 1998 Jun 21;192(4):455–65. \url{https://www.sciencedirect.com/science/article/pii/S0022519398906671}
}
\seealso{
\code{\link{kronauer}, \link{kronauer_default_parms}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
