% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_dreams_spindles}
\alias{download_dreams_spindles}
\title{Download the DREAMS Spindles database in the desired path. Check MD5 of the downloaded .rar file.
See the DREAMS Spindles database  website for more details: <http://www.tcts.fpms.ac.be/~devuyst/Databases/DatabaseSpindles/>}
\usage{
download_dreams_spindles(path)
}
\arguments{
\item{path}{Path}
}
\description{
Download the DREAMS Spindles database in the desired path. Check MD5 of the downloaded .rar file.
See the DREAMS Spindles database  website for more details: <http://www.tcts.fpms.ac.be/~devuyst/Databases/DatabaseSpindles/>
}
\references{
S. Devuyst et al., "Automatic Sleep Spindle Detection in Patients with Sleep Disorders," 2006 International Conference of the IEEE Engineering in Medicine and Biology Society, New York, NY, 2006, pp. 3883-3886. doi: 10.1109/IEMBS.2006.259298
}
