% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_dreams_subjects}
\alias{download_dreams_subjects}
\title{Download Subjects db}
\usage{
download_dreams_subjects(path)
}
\arguments{
\item{path}{target path.}
}
\description{
Download Subjects db
}
