% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philrob.R
\name{dPhilrob}
\alias{dPhilrob}
\title{Phillips and Robinson's model flow}
\usage{
dPhilrob(time, y, parms = philrob_default_parms())
}
\arguments{
\item{time}{The time (in h)}

\item{y}{The state}

\item{parms}{Model's parameters (optional, see \code{\link{philrob_default_parms}})}
}
\value{
The flow (right hand side of the differential equation)
}
\description{
Generates the flow (right hand side of the differential equation) for the Phillips and Robinson's model
}
\examples{
\dontrun{
t <- 0
y <- c(Vv = 1, Vm = 1, H = 1)
dy <- dPhilrob(t, y)
}
}
\references{
Phillips AJK, Robinson PA.
A Quantitative Model of Sleep-Wake Dynamics Based on the Physiology of the Brainstem Ascending Arousal System.
J Biol Rhythms. 2007 Apr 29;22(2):167–79. Available from: http://journals.sagepub.com/doi/10.1177/0748730406297512
}
\seealso{
\code{\link{philrob}, \link{philrob_default_parms}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
