% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philrob.R
\name{dSaturating_function}
\alias{dSaturating_function}
\title{Derivative of the saturating function}
\usage{
dSaturating_function(V, parms = philrob_default_parms())
}
\arguments{
\item{V}{voltage (in mV)}

\item{parms}{Model's parameters (optional, see \code{\link{philrob_default_parms}})}
}
\value{
The derivative of the saturation function
}
\description{
Derivative of the saturating function
}
\examples{
\dontrun{
dSaturating_function(1)
}
}
\references{
Phillips AJK, Robinson PA.
A Quantitative Model of Sleep-Wake Dynamics Based on the Physiology of the Brainstem Ascending Arousal System.
J Biol Rhythms. 2007 Apr 29;22(2):167–79. Available from: http://journals.sagepub.com/doi/10.1177/0748730406297512
}
\seealso{
\code{\link{philrob}, \link{saturating_function}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
