% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strogatz.R
\name{dStrogatz}
\alias{dStrogatz}
\title{Strogatz's model flow}
\usage{
dStrogatz(time, y, parms = strogatz_default_parms())
}
\arguments{
\item{time}{The time (in h)}

\item{y}{The state}

\item{parms}{Model's parameters (optional, see \code{\link{strogatz_default_parms}})}
}
\value{
The flow (right hand side of the differential equation)
}
\description{
Generates the flow (right hand side of the differential equation) for the Strogatz's model
}
\examples{
\dontrun{
t <- 0
y <- c(th1 = 0.1, th2 = 0.2)
dy <- dStrogatz(t, y)
}
}
\references{
Strogatz, S. H. (1987).
Human sleep and circadian rhythms: a simple model based on two coupled oscillators.
Journal of Mathematical Biology, 25(3), 327–347. \url{http://doi.org/10.1007/BF00276440}
}
\seealso{
\code{\link{strogatz}, \link{strogatz_default_parms}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
