% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Compute euclidean distance between two points}
\usage{
euclidean_distance(x, y)
}
\arguments{
\item{x}{numeric of length n giving X coordinates of a list of points}

\item{y}{numeric of same length as x giving Y coordinates of a list of points}
}
\value{
A numeric of length n giving the distance between point i and i-1.
Its initial value is NA
}
\description{
Compute euclidean distance between two points
}
\details{
Distance between 1,1 and 2,2 (should be square root of 2)
euclidean_distance(c(1,2), c(1,2))
NA 1.414214
}
