% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-functions.R
\name{generic_transition}
\alias{generic_transition}
\title{Compute a generic transition probability function}
\usage{
generic_transition(state0, state1)
}
\arguments{
\item{state0}{First state in the transition}

\item{state1}{Second state in the transition}
}
\value{
A function that computes P(state0|state1) in a sleep trace
}
\description{
A function that yields another function which computes the fraction of transitions
starting in state0 that end in state 1
Based on math from https://www.pnas.org/content/117/18/10024
}
\details{
The resulting function takes a trace of binary states.

If using the column asleep from a behavr object, 1 encodes asleep and 0 awake
}
