% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{get.hypnogram}
\alias{get.hypnogram}
\title{Filter an events dataframe to keep hypnogram.}
\usage{
get.hypnogram(events)
}
\arguments{
\item{events}{events dataframe.}
}
\value{
hypngram dataframe.
}
\description{
Filter an events dataframe to keep hypnogram.
}
