% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draft_order.R
\name{get_draft_order}
\alias{get_draft_order}
\title{Get Draft Order from Sleeper League}
\usage{
get_draft_order(leauge_id = "1069393183089586176")
}
\arguments{
\item{leauge_id}{A character string representing the ID of the Sleeper fantasy league. Defaults to "1069393183089586176".}
}
\value{
A data frame containing the draft order, team names, the reason for their placement (playoff or non-playoff team), and a description of the reason.
}
\description{
This function retrieves and calculates the draft order for a specified Sleeper fantasy league. It integrates playoff and non-playoff team performances to generate the draft order based on team placements and total points scored.
}
\details{
The function performs the following steps:
\itemize{
\item Retrieves aggregated points data for the league.
\item Fetches playoff bracket information from the Sleeper API.
\item Processes playoff results to determine team placements.
\item Calculates placements for non-playoff teams based on total points scored.
\item Combines playoff and non-playoff results into a final draft order.
}
}
\examples{
# Example usage:
draft_order <- get_draft_order("123456789012345678")
print(draft_order)

}
\seealso{
\code{\link[sleepr:create_aggregated_points_df]{create_aggregated_points_df}}
}
