% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_n1_minutes}
\alias{get_n1_minutes}
\title{Get total duration of N1 sleep in minutes.}
\usage{
get_n1_minutes(hypnogram)
}
\arguments{
\item{hypnogram}{Hypnogram dataframe.}
}
\value{
total duration of N1 sleep in minutes.
}
\description{
Get total duration of N1 sleep in minutes.
}
