% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_overlapping_duration}
\alias{get_overlapping_duration}
\title{get_overlapping_duration}
\usage{
get_overlapping_duration(label, stages, events)
}
\arguments{
\item{label}{label}

\item{stages}{stages}

\item{events}{events}
}
\description{
get_overlapping_duration
}
