% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_overlapping_events}
\alias{get_overlapping_events}
\title{Filter x events overlapping y events.}
\usage{
get_overlapping_events(e, x, y)
}
\arguments{
\item{e}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}

\item{x}{Events labels overlapping y events.}

\item{y}{Events labels overlapped by x events.}
}
\value{
Dataframe of x events overlapping y events.
}
\description{
Filter x events overlapping y events.
}
