% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_rem_minutes}
\alias{get_rem_minutes}
\title{Get total duration of REM sleep in minutes.}
\usage{
get_rem_minutes(events)
}
\arguments{
\item{events}{events dataframe.}
}
\value{
total duration of REM sleep in minutes.
}
\description{
Get total duration of REM sleep in minutes.
}
