% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekly_points.R
\name{get_weekly_points}
\alias{get_weekly_points}
\title{Get Weekly Points}
\usage{
get_weekly_points(
  league_id = "1124848060283768832",
  nfl_start_dt = "2024-09-05",
  player_df = NA
)
}
\arguments{
\item{league_id}{A string representing the ID of the fantasy football league (default: "1124848060283768832").}

\item{nfl_start_dt}{A string representing the start date of the NFL season (default: "2024-09-05").}

\item{player_df}{A dataframe containing player metadata. If not provided, the function dynamically retrieves player data using \code{get_player_df}.}
}
\value{
A list with two elements:
\describe{
\item{\code{full_points_df}}{A dataframe summarizing weekly points for each roster, including starter points, bench points, and win/loss outcomes.}
\item{\code{long_points_df}}{A detailed dataframe containing individual player points for each week and roster.}
}
}
\description{
Retrieves and calculates weekly points for a fantasy football league, including starter points, bench points, and optimized starter points.
}
\details{
This function fetches weekly matchup data from the Sleeper API for a specified league, processes player and team data,
and computes various point summaries for each week. It leverages the \code{create_season_weekly_df},
\code{create_starter_df}, and \code{optimize_starter_points} functions to organize and optimize the data.
}
\examples{
\dontrun{
# Fetch weekly points for a league
weekly_points <- get_weekly_points(
  league_id = "1124848060283768832",
  nfl_start_dt = "2024-09-05"
)
full_points_df <- weekly_points$full_points_df
long_points_df <- weekly_points$long_points_df
print(full_points_df)
print(long_points_df)
}
}
