% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming-index.R
\name{hamming_index}
\alias{hamming_index}
\title{Compute hamming index with a sliding window}
\usage{
hamming_index(asleep_sequence)
}
\arguments{
\item{asleep_sequence}{A logical vector stating the sequence of asleep/awake states}
}
\value{
A double (numeric) which results from taking the hamming index
and dividing by the length of the sequence
}
\description{
Takes a sequence of asleep/awake states in logical format (TRUE-> asleep)
and computes the distance to the closes fully consolidated sequence of equal
length and proportion of asleep/awake states
}
\details{
The hamming index of a sleep run i the minimal number pairs of positions
that need to be swapped to get to a continuous run of asleep states i.e.
a run with only one sleeping bout
}
