% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{hypnogram_band_powers}
\alias{hypnogram_band_powers}
\title{Compute bands powers from a full hypnogram.}
\usage{
hypnogram_band_powers(record, channel, bands = list(delta = c(0.5, 3.5),
  theta = c(3.5, 8), alpha = c(8, 12), beta = c(12, 30), gamma1 = c(30,
  40)), normalize = c(4, 40))
}
\arguments{
\item{record}{record.}

\item{channel}{channel to split and hyp.}

\item{bands}{bands. Default "delta","theta","alpha","beta", "gamma1", "denominator", "broadband".}

\item{normalize}{normalize.}
}
\value{
a df.
}
\description{
Compute bands powers from a full hypnogram.
}
