% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{hypnogram_eemd}
\alias{hypnogram_eemd}
\title{Compute EEMD mean,var,skew,kur}
\usage{
hypnogram_eemd(record, channel, num_imfs = 8, labels = c("N3", "N2",
  "N1", "REM", "AWA"), butter = FALSE)
}
\arguments{
\item{record}{record.}

\item{channel}{channel to split and hyp.}

\item{num_imfs}{numinfs}

\item{labels}{lablssl}

\item{butter}{butter}
}
\value{
a df.
}
\description{
Compute EEMD mean,var,skew,kur
}
