% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronauer.R
\name{kronauer}
\alias{kronauer}
\title{Solve Kronauer's 1990 model}
\usage{
kronauer(ts, y0, parms = kronauer_default_parms(), ...)
}
\arguments{
\item{ts}{Vector of times (in h)}

\item{y0}{Initial condition}

\item{parms}{Model parameters (optional, see \code{\link{kronauer_default_parms}})}

\item{...}{Additional arguments passed to the \code{\link{ode}} integrator}
}
\value{
Results of the simulation, including times and states
}
\description{
Solves the Kronauer's 1990 model (equations 2 and 3 in the reference) for the given times, initial condition and parameters
}
\examples{
y0 <- c(x = 1, xc = 0)
nDays <- 5
ts <- seq(0, nDays*24, length.out = nDays*24*20)
I <- function(t) { 9500 + 0*t }
sol <- kronauer(ts, y0, I)
}
\references{
Jewett ME, Kronauer RE.
Refinement of Limit Cycle Oscillator Model of the Effects of Light on the Human Circadian Pacemaker.
J Theor Biol. 1998 Jun 21;192(4):455–65. \url{https://www.sciencedirect.com/science/article/pii/S0022519398906671}
}
\seealso{
\code{\link{dKronauer}, \link{kronauer_default_parms}, \link{ode}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
