% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronauer.R
\name{kronauer_default_parms}
\alias{kronauer_default_parms}
\title{Default parameters of Kronauer's 1990 model}
\usage{
kronauer_default_parms()
}
\value{
The default parameters for Kronauer's 1990 model
}
\description{
Loads the parameters used in Kronauer's 1990 model
}
\examples{
parms <- kronauer_default_parms()
}
\references{
Jewett ME, Kronauer RE.
Refinement of Limit Cycle Oscillator Model of the Effects of Light on the Human Circadian Pacemaker.
J Theor Biol. 1998 Jun 21;192(4):455–65. \url{https://www.sciencedirect.com/science/article/pii/S0022519398906671}
}
\seealso{
\code{\link{kronauer}, \link{dKronauer}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
