% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{lissajous_figure}
\alias{lissajous_figure}
\title{Lissajous figure}
\usage{
lissajous_figure(times, ys, wx = 2 * pi/24, Ax = 1, alphax = 0, ...)
}
\arguments{
\item{times}{Times vector}

\item{ys}{State vector}

\item{wx}{(Optional) Reference frequency. Default: 2*pi/24}

\item{Ax}{(Optional) Reference amplitude. Default: 1}

\item{alphax}{(Optional) Reference phase delay. Default: 0}

\item{...}{(Optional) Graphical parameters}
}
\value{
Plots the Lissajous figure
}
\description{
Lissajous figure
}
\examples{
# Simulate a solution
nDays <- 8
ts <- seq(0, nDays * 24, length.out = nDays * 24 * 20)
y0 <- c(Vv = -12.6404, Vm = 0.8997, H = 12.5731) # This point is already close to the attractor
sol <- philrob(ts, y0, method = 'lsode')
# Create the figure
lissajous_figure(ts, sol$H)
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
