% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ma_duration}
\alias{ma_duration}
\title{Micro-arousals duration}
\usage{
ma_duration(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Micro-arousals duration
}
