% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ma_n2_duration}
\alias{ma_n2_duration}
\title{Micro-arousals n2 duration}
\usage{
ma_n2_duration(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Micro-arousals n2 duration
}
