% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ma_n3_count}
\alias{ma_n3_count}
\title{Micro-arousals n3 count}
\usage{
ma_n3_count(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Micro-arousals n3 count
}
