% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ma_stats}
\alias{ma_stats}
\title{Get Micro-Arousals events related stats in a named vector.}
\usage{
ma_stats(e)
}
\arguments{
\item{e}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\value{
A named vector.
}
\description{
Get Micro-Arousals events related stats in a named vector.
}
