% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_annotation.R
\name{median_movement_detector}
\alias{median_movement_detector}
\title{Find the median distance traversed by the animal}
\usage{
median_movement_detector(data, time_window_length = 10, ...)
}
\arguments{
\item{data}{\link{data.table} containing behavioural variable from or one multiple animals.
When it has a key, unique values, are assumed to represent unique individuals (e.g. in a \link{behavr} table).
Otherwise, it analysis the data as coming from a single animal. \code{data} must have a column \code{t} representing time.}

\item{time_window_length}{number of seconds to be used by the motion classifier.
This corresponds to the sampling period of the output data.}

\item{...}{extra arguments to be passed to \code{motion_classifier_FUN}.}
}
\description{
Find the median distance traversed by the animal
}
