% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_annotation.R
\name{movement_detector_enclosed}
\alias{movement_detector_enclosed}
\alias{max_movement_detector}
\alias{median_movement_detector}
\alias{sum_movement_detector}
\title{Generic function to aggregate movement with some statistic}
\usage{
movement_detector_enclosed(func, feature, statistic, score, preproc_FUN = NULL)

max_movement_detector(data, time_window_length = 10, ...)

median_movement_detector(data, time_window_length = 10, ...)

sum_movement_detector(data, time_window_length = 10, ...)
}
\arguments{
\item{func}{Aggregating function (max, min, median, mean, etc)}

\item{feature}{Name of a column in the sqlite3 file e.g. body_movement}

\item{statistic}{Name of the column resulting from aggregation e.g. max_movement}

\item{score}{Name of the column providing a score i.e. category to the statistic e.g. micromovement
score is usually a binary variable i.e. TRUE/FALSE}

\item{preproc_FUN}{Optional, function to preprocess the input before computing the feature
(if the data needs some transformation like reverting xy_dist_log10x1000 back to a distance)}

\item{time_window_length}{Size of non overlapping time bins, in seconds}

\item{threshold}{If the statistic is greater than this value, the score is TRUE, and 0 otherwise}
}
\description{
Generic function to aggregate movement with some statistic
}
