% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n1_duration}
\alias{n1_duration}
\title{Sums up N1 stages duration from events dataframe to get total N1 duration in minutes.}
\usage{
n1_duration(events)
}
\arguments{
\item{events}{Hypnogram dataframe.}
}
\value{
total duration of N1 sleep in minutes.
}
\description{
Sums up N1 stages duration from events dataframe to get total N1 duration in minutes.
}
