% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n1_tts}
\alias{n1_tts}
\title{Divides N1 duration by TTS duration from an events dataframe.}
\usage{
n1_tts(hypnogram)
}
\arguments{
\item{events}{Events dataframe. Must contain begin, end and events.}
}
\value{
N1 over TTS durations ratio.
}
\description{
Divides N1 duration by TTS duration from an events dataframe.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N3","N1")
n1_tts(events)
}
