% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n2_minutes}
\alias{n2_minutes}
\title{Get total duration of N2 sleep in minutes.}
\usage{
n2_minutes(hypnogram)
}
\arguments{
\item{hypnogram}{Hypnogram dataframe.}
}
\value{
total duration of N2 sleep in minutes.
}
\description{
Get total duration of N2 sleep in minutes.
}
