% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n3_tts}
\alias{n3_tts}
\title{Divides N3 duration by TTS duration from an events dataframe.}
\usage{
n3_tts(events)
}
\arguments{
\item{events}{Events dataframe. Must contain begin, end and events.}
}
\value{
N3 over TTS durations ratio.
}
\description{
Divides N3 duration by TTS duration from an events dataframe.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N3","REM")
n3_tts(events)
}
