% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-functions.R
\name{p_doze}
\alias{p_doze}
\title{Compute P(D|W), the probability that the fly changes from a wake to a doze state}
\usage{
p_doze(asleep_sequence)
}
\description{
Takes as input a sequence of binary states (1/0 or T/F) assumming 1 is asleep or doze state
}
