% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-functions.R
\name{p_wake}
\alias{p_wake}
\title{Compute P(W|D), the probability that the fly changes from a doze to a wake state}
\usage{
p_wake(asleep_sequence)
}
\description{
Takes as input a sequence of binary states (1/0 or T/F) assumming 1 is asleep or doze state
}
