% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philrob.R
\name{philrob_default_forcing}
\alias{philrob_default_forcing}
\title{Default forcing function}
\usage{
philrob_default_forcing(parms = philrob_default_parms())
}
\arguments{
\item{parms}{Parameters (optional)}
}
\value{
The standard offset + cosine forcing function
}
\description{
Default forcing function
}
\examples{
\dontrun{
C <- function(t) { philrob_default_forcing(t) }
t <- seq(0, 3, length.out = 10)
}
}
\references{
Phillips AJK, Robinson PA.
A Quantitative Model of Sleep-Wake Dynamics Based on the Physiology of the Brainstem Ascending Arousal System.
J Biol Rhythms. 2007 Apr 29;22(2):167–79. Available from: http://journals.sagepub.com/doi/10.1177/0748730406297512
}
\seealso{
\code{\link{philrob}, \link{dPhilrob}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
