% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philrob.R
\name{philrob_simplified}
\alias{philrob_simplified}
\title{Solve simplified Phillip and Robinson's model}
\usage{
philrob_simplified(ts, y0, drive, parms = philrob_default_parms(), ...)
}
\arguments{
\item{ts}{Vector of times (in h)}

\item{y0}{Initial condition}

\item{drive}{The external drive, as a function of t (in h)}

\item{parms}{Model parameters (optional, see \code{\link{philrob_default_parms}})}

\item{...}{Additional arguments passed to the \code{\link{ode}} integrator}
}
\value{
Results of the simulation, including times, states and asleep/awake status
}
\description{
Solves the Phillip and Robinson's model for the given times, initial condition and parameters
}
\examples{
y0 <- c(Vv = -13, Vm = 1)
nDays <- 30
ts <- seq(0, nDays*24, length.out = nDays*24*20)
D <- function(t) { 1 + 0.5*cos(2*pi*t/24) }
sol <- philrob_simplified(ts, y0, drive = D(ts))
}
\references{
Phillips AJK, Robinson PA.
A Quantitative Model of Sleep-Wake Dynamics Based on the Physiology of the Brainstem Ascending Arousal System.
J Biol Rhythms. 2007 Apr 29;22(2):167–79. Available from: http://journals.sagepub.com/doi/10.1177/0748730406297512
}
\seealso{
\code{\link{dPhilrob_simplified}, \link{philrob_default_parms}, \link{ode}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
