% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{philrobPlot}
\alias{philrobPlot}
\title{Plots the time series generated with philrob}
\usage{
philrobPlot(sol)
}
\arguments{
\item{sol}{Solution of a philrob simulation}
}
\value{
The simulation's plot
}
\description{
Plots the time series generated with philrob
}
\examples{
# Generate a solution
y0 <- c(Vv = -13, Vm = 1, H = 10)
nDays <- 30
ts <- seq(0, nDays*24, length.out = nDays*24*20)
sol <- philrob(ts, y0)
# Plot it
philrobPlot(sol)
}
\references{
Phillips AJK, Robinson PA.
A Quantitative Model of Sleep-Wake Dynamics Based on the Physiology of the Brainstem Ascending Arousal System.
J Biol Rhythms. 2007 Apr 29;22(2):167–79. Available from: http://journals.sagepub.com/doi/10.1177/0748730406297512
}
\seealso{
\code{\link{philrob}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
