% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{plot.hypnogram}
\alias{plot.hypnogram}
\title{Draw a histogram with ggplot.}
\usage{
\method{plot}{hypnogram}(events)
}
\arguments{
\item{events}{events dataframe.}
}
\value{
a ggplot.
}
\description{
Draw a histogram with ggplot.
}
