% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{plot_hypnogram}
\alias{plot_hypnogram}
\title{Draw a histogram with ggplot2.}
\usage{
plot_hypnogram(events)
}
\arguments{
\item{events}{events dataframe.}
}
\value{
a ggplot.
}
\description{
Draw a histogram with ggplot2.
}
