% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{plot_spectrogram}
\alias{plot_spectrogram}
\title{Get spectrogram}
\usage{
plot_spectrogram(signal, sRate, maxfreq = 25, windowlength = 2000,
  show = FALSE)
}
\arguments{
\item{signal}{signal.}

\item{sRate}{sRate.}

\item{maxfreq}{maxfreq.}

\item{windowlength}{windowlength.}

\item{show}{show.}
}
\value{
list.
}
\description{
Get spectrogram
}
