% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion_detectors.R
\name{prepare_data_for_motion_detector}
\alias{prepare_data_for_motion_detector}
\title{Copy needed columns, and add t_round var for downsampling}
\usage{
prepare_data_for_motion_detector(
  data,
  needed_columns,
  time_window_length,
  optional_columns = NULL,
  curate = TRUE
)
}
\arguments{
\item{data}{\link{data.table} containing behavioural variable from or one multiple animals.
When it has a key, unique values, are assumed to represent unique individuals (e.g. in a \link{behavr} table).
Otherwise, it analysis the data as coming from a single animal. \code{data} must have a column \code{t} representing time.}

\item{needed_columns}{Columns that must be present in the data}

\item{time_window_length}{number of seconds to be used by the motion classifier.
This corresponds to the sampling period of the output data.}

\item{optional_columns}{Optional, columns that can be used if available}

\item{curate}{If true, remove sparse ROI datasets (probably spurious fly detections)}
}
\description{
Copy needed columns, and add t_round var for downsampling
}
