% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{rasterPlot}
\alias{rasterPlot}
\title{Raster plot / somnogram}
\usage{
rasterPlot(sol, dailySamples = 480)
}
\arguments{
\item{sol}{Simulation (output of \code{\link{strogatz}})}

\item{dailySamples}{Number of samples per day (optional)}
}
\value{
Plots the raster plot / somnogram
}
\description{
Raster plot / somnogram
}
\examples{
# Simulate a solution
y0 <- c(th1 = 0.1, th2 = 0.05)
nDays <- 60.2
ts <- seq(0, nDays*24, length.out=nDays*24*20)
sol <- strogatz(ts, y0)
# Create raster plot
rasterPlot(sol)
}
