% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read.events.noxturnal}
\alias{read.events.noxturnal}
\title{Read a Noxturnal events file (Unicode CSV format)}
\usage{
read.events.noxturnal(path)
}
\arguments{
\item{path}{Noxturnal events file path.}
}
\value{
A dataframe of scored events.
}
\description{
Read a Noxturnal events file (Unicode CSV format)
}
