% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.mdf}
\alias{read.mdf}
\title{Read a MDF}
\usage{
read.mdf(mdfPath, channels = c())
}
\arguments{
\item{mdfPath}{MDF path.}

\item{channels}{Channels to read.}
}
\value{
A (large) list containing signals and metadata.
}
\description{
Read a MDF
}
