% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_all_events_isruc}
\alias{read_all_events_isruc}
\title{Read scoring from the ISRUC-Sleep datasets.}
\usage{
read_all_events_isruc(dir)
}
\arguments{
\item{dir}{ISRUC target.}
}
\value{
A dataframe of scored events.
}
\description{
Read scoring from the ISRUC-Sleep datasets.
}
