% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_events_dreams_subjects}
\alias{read_events_dreams_subjects}
\title{Read Subjects DB AASM scoring}
\usage{
read_events_dreams_subjects(record_id, path)
}
\arguments{
\item{record_id}{record_id}

\item{path}{path.}
}
\value{
A dataframe of scored events.
}
\description{
Read Subjects DB AASM scoring
}
