% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_events_isruc}
\alias{read_events_isruc}
\title{Read scoring from the ISRUC-Sleep datasets.}
\usage{
read_events_isruc(dir, scoringNum = 1)
}
\arguments{
\item{dir}{character, ISRUC record directory.}

\item{scoringNum}{Scoring number in database.}
}
\value{
A dataframe of scored stages.
}
\description{
Read a XLSX sleep stages scoring from  a ISRUC-Sleep dataset record directory.
}
\references{
Sirvan Khalighi, Teresa Sousa, Jose Moutinho Santos, Urbano Nunes, ISRUC-Sleep: a comprehensive public dataset for sleep researchers, Computer Methods and Programs in Biomedicine, Elsevier, 2015
}
