% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_isruc_metadata}
\alias{read_isruc_metadata}
\title{Read ISRUC metadata}
\usage{
read_isruc_metadata(target)
}
\arguments{
\item{target}{ISRUC records directory.}
}
\value{
A dataframe of metadata.
}
\description{
Read ISRUC metadata
}
