% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mdf}
\alias{read_mdf}
\title{Read a MDF}
\usage{
read_mdf(mdfPath, channels = c(NA), metadata = TRUE)
}
\arguments{
\item{mdfPath}{[String] MDF path.}

\item{channels}{[String] Channels to read.}

\item{metadata}{[Boolean] Read or not the metadata.}
}
\value{
A (large) list.
}
\description{
Read a MDF
}
