% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rem_avg_duration}
\alias{rem_avg_duration}
\title{REM avg duration}
\usage{
rem_avg_duration(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
REM avg duration
}
