% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rem_count}
\alias{rem_count}
\title{REM count}
\usage{
rem_count(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
REM count
}
