% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rem_duration}
\alias{rem_duration}
\title{Sums up REM stages duration from an events dataframe to get total REM duration in minutes.}
\usage{
rem_duration(events)
}
\arguments{
\item{events}{events dataframe. Must have begin, end and event variables.}
}
\value{
Total duration of REM sleep in minutes.
}
\description{
Sums up REM stages duration from an events dataframe to get total REM duration in minutes.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("REM","REM")
rem_duration(events)
}
