% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rem_index}
\alias{rem_index}
\title{REM index}
\usage{
rem_index(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
REM index
}
