% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rem_latency}
\alias{rem_latency}
\title{REM Latency in minutes.}
\usage{
rem_latency(events)
}
\arguments{
\item{events}{Events dataframe. Must contain begin, end and events.}
}
\value{
Sleep latency.
}
\description{
Substracts the start time of the first REM epoch to the start of the sleep onset.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("AWA","REM")
rem_latency(events)

events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("REM","REM")
rem_latency(events)

events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N1","REM")
rem_latency(events)
}
