% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rem_minutes}
\alias{rem_minutes}
\title{Get total duration of REM sleep in minutes.}
\usage{
rem_minutes(hypnogram)
}
\arguments{
\item{hypnogram}{Hypnogram dataframe.}
}
\value{
total duration of REM sleep in minutes.
}
\description{
Get total duration of REM sleep in minutes.
}
