% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rem_tts}
\alias{rem_tts}
\title{REM over TTS ratio.}
\usage{
rem_tts(events)
}
\arguments{
\item{events}{Events dataframe. Must contain begin, end and events.}
}
\value{
REM over TTS durations ratio.
}
\description{
Divides REM duration by TTS duration from an events dataframe.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N1","REM")
rem_tts(events)
}
