% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{reshape_as_raster}
\alias{reshape_as_raster}
\title{Reshape as raster vector}
\usage{
reshape_as_raster(sol, dailySamples = 480)
}
\arguments{
\item{sol}{Simulation (output of \code{\link{strogatz}})}

\item{dailySamples}{Number of samples per day (optional)}
}
\value{
The asleep vector reshaped as a raster vector
}
\description{
Reshape as raster vector
}
\examples{
\dontrun{
# Simulate a solution
y0 <- c(th1 = 0.1, th2 = 0.05)
nDays <- 60.2
ts <- seq(0, nDays*24, length.out=nDays*24*20)
sol <- strogatz(ts, y0)
# Get raster sleep vector
asleep_raster <- reshape_as_raster(sol)
}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
