% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{resp_stats}
\alias{resp_stats}
\title{Get Apnea and Hypopnea events related stats in a named vector.}
\usage{
resp_stats(e, ss = c("N1", "N2", "N3", "REM"), l = c("A. Obstructive",
  paste0("Hypopn", "é", "e")))
}
\arguments{
\item{e}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}

\item{ss}{sleep labels.}

\item{l}{label.}
}
\value{
A named vector.
}
\description{
Get Apnea and Hypopnea events related stats in a named vector.
}
