% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philrob.R
\name{saturating_function}
\alias{saturating_function}
\title{Saturating function}
\usage{
saturating_function(V, parms = philrob_default_parms())
}
\arguments{
\item{V}{voltage (in mV)}

\item{parms}{Model's parameters (optional, see \code{\link{philrob_default_parms}})}
}
\value{
The saturated effect of the voltage
}
\description{
Saturating function
}
\examples{
\dontrun{
saturating_function(1)
}
}
\references{
Phillips AJK, Robinson PA.
A Quantitative Model of Sleep-Wake Dynamics Based on the Physiology of the Brainstem Ascending Arousal System.
J Biol Rhythms. 2007 Apr 29;22(2):167–79. Available from: http://journals.sagepub.com/doi/10.1177/0748730406297512
}
\seealso{
\code{\link{philrob}, \link{philrob_default_parms}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
