% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep-annotation.R
\name{sleep_annotation_closure}
\alias{sleep_annotation_closure}
\title{Score sleep behaviour from immobility}
\usage{
sleep_annotation_closure(velocity_correction_coef)
}
\description{
This function creates a sleep_annotation function with a custom
velocity correction coefficient i.e. the user can change the default of
0.003.
}
