% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sleep_efficiency}
\alias{sleep_efficiency}
\title{Divides the Time To Sleep (TTS) by the Total Sleep Period (TSP) to get the sleep efficiency ratio.}
\usage{
sleep_efficiency(events)
}
\arguments{
\item{events}{Events dataframe. Must contain begin, end and events.}
}
\value{
sleep efficiency ratio.
}
\description{
Divides the Time To Sleep (TTS) by the Total Sleep Period (TSP) to get the sleep efficiency ratio.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N3","AWA")
sleep_efficiency(events)
}
