% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sleep_latency}
\alias{sleep_latency}
\title{Substracts the start time of the first sleep epoch (N1,N2,N3 or REM) to the start of the hypnogram.}
\usage{
sleep_latency(events)
}
\arguments{
\item{events}{Events dataframe. Must contain begin, end and events.}
}
\value{
Sleep latency.
}
\description{
Substracts the start time of the first sleep epoch (N1,N2,N3 or REM) to the start of the hypnogram.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("AWA","N3")
sleep_latency(events)
}
