\name{sleepR-package}
\alias{sleepR-package}
\alias{sleepR}
\docType{package}
\title{
Analysis of sleeping time
}
\description{
Process data collected with Android app 'Sleepbot'
}
\details{
\tabular{ll}{
Package: \tab sleepR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-12-14\cr
License: \tab GPL-2\cr
}

}
\author{
Giovanni Bonafe' <midable@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
