% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{snoring_duration_pct}
\alias{snoring_duration_pct}
\title{Total duration of snorings during TTS over TTS duration.}
\usage{
snoring_duration_pct(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\value{
Total duration of snorings during TTS over TTS.
}
\description{
\code{snoring_duration} divides the durations of snorings during TTS (N1,N2,N3,REM) over the TTS duration.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830,1536967810),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860,1536967820), origin = "1970-01-01")
events$event = c("N3","N3","Train de ronflements")
snoring_duration_pct(events)
}
