% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{snoring_index}
\alias{snoring_index}
\title{Snoring index}
\usage{
snoring_index(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\value{
Snoring index
}
\description{
\code{snoring_index} return the number of snorings occuring during TTS over TTS in hours.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830,1536967810),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860,1536967820), origin = "1970-01-01")
events$event = c("N3","N3","Train de ronflements")
snoring_index(events)
}
