% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{split_epochs}
\alias{split_epochs}
\title{Split signal into epochs}
\usage{
split_epochs(signal, sRate, duration = 30)
}
\arguments{
\item{signal}{signal.}

\item{sRate}{sRate sig.}

\item{duration}{in seconds.}
}
\value{
list.
}
\description{
Split signal into epochs
}
