% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{split_signal}
\alias{split_signal}
\title{Split signal using hypnogram events.}
\usage{
split_signal(signal, hypnogram, sRate)
}
\arguments{
\item{signal}{Vector containing signal.}

\item{hypnogram}{hypnogram df.}

\item{sRate}{Signal sample rate.}
}
\value{
list.
}
\description{
Split signal using hypnogram events.
}
