% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strogatz.R
\name{strogatz_default_parms}
\alias{strogatz_default_parms}
\title{Default parameters of Strogatz's model}
\usage{
strogatz_default_parms()
}
\value{
The default parameters for Strogatz's model
}
\description{
Loads the parameters used in Strogatz's model
}
\details{
Strogatz, S. H. (1987).
Human sleep and circadian rhythms: a simple model based on two coupled oscillators.
Journal of Mathematical Biology, 25(3), 327–347. \url{http://doi.org/10.1007/BF00276440}
}
\examples{
parms <- strogatz_default_parms()
}
\references{
Strogatz, S. H. (1987).
Human sleep and circadian rhythms: a simple model based on two coupled oscillators.
Journal of Mathematical Biology, 25(3), 327–347. \url{http://doi.org/10.1007/BF00276440}
}
\seealso{
\code{\link{strogatz}, \link{dStrogatz}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
