% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tm_stats}
\alias{tm_stats}
\title{Flattens a transition matrix to a named vector.}
\usage{
tm_stats(tm)
}
\arguments{
\item{tm}{A transition matrix.}
}
\value{
A named vector.
}
\description{
Flattens a transition matrix to a named vector to ensure stats functions consistency.
}
