% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tsp}
\alias{tsp}
\title{D}
\usage{
tsp(events)
}
\arguments{
\item{events}{Events dataframe. Must contain begin, end and events.}
}
\value{
N1 over TTS durations ratio.
}
\description{
D
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N3","N1")
tsp(events)
}
