% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tts}
\alias{tts}
\title{Time To Sleep in minutes.}
\usage{
tts(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\value{
Time To Sleep (N1+N2+N3+REM durations) in minutes.
}
\description{
\code{tts} (Time To Sleep) sums up REM, N1, N2 and N3 stages duration from an events dataframe to get Time To Sleep duration in minutes.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N1","REM")
tts(events)
}
