% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tts_pos_nonback}
\alias{tts_pos_nonback}
\title{TTS duration in non-back position in minutes.}
\usage{
tts_pos_nonback(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\value{
TTS duration in non-back position.
}
\description{
\code{tts_pos_nonback} computes the total time in non-back position during TTS in minutes.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830,1536967810),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860,1536967820), origin = "1970-01-01")
events$event = c("N3","N3","back")
tts_pos_nonback(events)
}
