% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_annotation.R
\name{velocity_avg_enclosed}
\alias{velocity_avg_enclosed}
\alias{velocity_avg}
\title{Compute velocity aggregates using xy_dist_log10x1000}
\usage{
velocity_avg_enclosed(data, time_window_length = 10)

velocity_avg(data, time_window_length = 10, ...)
}
\description{
Compute velocity aggregates using xy_dist_log10x1000
}
