% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{waso}
\alias{waso}
\title{Wake After Sleep Onset in minutes.}
\usage{
waso(events)
}
\arguments{
\item{events}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\value{
Wake After Sleep Onset in minutes.
}
\description{
\code{waso} substracts Time To Sleep (TTS) and sleep latency to the Total Sleep Period (TSP) to get the total Wake After Sleep Onset in minutes.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N2","AWA")
waso(events)
}
