% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_channel}
\alias{write_channel}
\title{Write a channel, MDF style.}
\usage{
write_channel(channel, signals, headers, mdfPath)
}
\arguments{
\item{channel}{Channel name.}

\item{signals}{signals list.}

\item{headers}{headers.}

\item{mdfPath}{mdf path}
}
\description{
Write a channel, MDF style.
}
