% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_mdf}
\alias{write_mdf}
\title{Write a MDF from an EDF file. MDF directory is erased if it already exists.}
\usage{
write_mdf(edfPath, mdfPath, channels = c(NA), events = c())
}
\arguments{
\item{edfPath}{EDF file path.}

\item{mdfPath}{MDF path.}

\item{channels}{Channels to write.}

\item{events}{Events dataframe to write.}
}
\description{
Write a MDF from an EDF file. MDF directory is erased if it already exists.
}
