% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{bands_power}
\alias{bands_power}
\title{Compute bands powers.}
\usage{
bands_power(x, Fs, bands, normalize, butter = FALSE)
}
\arguments{
\item{x}{signal vector.}

\item{Fs}{Frequency.}

\item{bands}{bands.}

\item{normalize}{normalize.}

\item{butter}{butter order or false.}
}
\value{
a named vector with all bands powers.
}
\description{
Compute bands powers.
}
