% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout-analysis.R
\name{bout_analysis}
\alias{bout_analysis}
\title{Finds "bouts" in categorical time series.}
\usage{
bout_analysis(var, data)
}
\arguments{
\item{var}{Column variable to use in \code{data}}

\item{data}{\code{data.table} representing a behaviour of multiple animals.}
}
\value{
A \code{data.table} where \strong{each row is a bout} and with the columns:
\itemize{
\item \code{start_time} The onset of this bout (second, same reference as \code{t})
\item \code{length} The bout duration (second)
\item A column with the same name as \code{var} that describe the value of var for this whole bout.
}
}
\description{
This function is used to find contiguous regions of unique value
in a -- potentially irregular -- univariate categorical time series.
}
\details{
Bout analysis will be performed by biological individual.
Individuals are defined by the \code{data.table} key.
}
\examples{
set.seed(1)
# 1000 points the first 500 points should have higher chance to be 1 than the last 500:
y_var <- round(c(runif(500,0,1),
                   runif(500,0,0.75)))
# first 500 point are for individual "A", next 500 points are for "B":
dt <- data.table( y = y_var,
             t = rep(1:500,2)*12,
             id = rep(c("A","B"),each=500),key="id")
print(dt)
bout_dt <- boutAnalysis(y,dt)
print(bout_dt)
# average duration and number of bouts for both individuals and
# bouts of y=0 or y=1
summary_dt <- bout_dt[,
        .(n=.N,
          mean_duration=mean(length))
        ,by=c(key(bout_dt),"y")]
print(summary_dt)
}
\seealso{
\itemize{
\item Tutorial on anlysis of sleep architecture \url{http://gilestrolab.github.io/rethomics/tutorial/sleep_quantification.html#sleep-architecture} TODO
\item \link{rle} run length encoding function, on which this analysis is based
}
}
