% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_all_stats}
\alias{compute_all_stats}
\title{Call and compile all statistics functions in a dataframe.}
\usage{
compute_all_stats(records, eeg_channels = c("C3-A2", "EEG Fpz-Cz"),
  metadata = TRUE)
}
\arguments{
\item{records}{record paths.}

\item{eeg_channels}{potential EEG channel names.}

\item{metadata}{read metadata or not.}
}
\value{
df.
}
\description{
Call and compile all statistics functions in a dataframe.
}
