% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_annotation.R
\name{custom_annotation_wrapper}
\alias{custom_annotation_wrapper}
\title{Custom annotation from the dt_raw file}
\usage{
custom_annotation_wrapper(custom_function)
}
\arguments{
\item{custom_function}{function used to produce the custom annotation}

\item{data}{\link{data.table} containing behavioural variable from or one multiple animals.
When it has a key, unique values, are assumed to represent unique individuals (e.g. in a \link{behavr} table).
Otherwise, it analysis the data as coming from a single animal. \code{data} must have a column \code{t} representing time.}

\item{time_window_length}{number of seconds to be used by the motion classifier.
This corresponds to the sampling period of the output data.}

\item{...}{extra arguments to be passed to \code{custom_function}.}
}
\value{
a \link{behavr} table similar to \code{data} with additional variables/annotations.
The resulting data will only have one data point every \code{time_window_length} seconds.
}
\description{
This function gives aggregates a variable of interest in a custom way
All datapoints in every time_window_length seconds is aggregated into a single datapoint
}
\details{
The default \code{time_window_length} is 300 seconds -- it is also known as the "5-minute rule".
}
\seealso{

}
