% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_annotation.R
\name{movement_detector_enclosed}
\alias{movement_detector_enclosed}
\alias{custom_annotation_wrapper}
\alias{max_movement_detector}
\alias{median_movement_detector}
\alias{sum_movement_detector}
\title{Generic function to aggregate movement with some statistic}
\usage{
movement_detector_enclosed(func, feature, statistic, score, preproc_FUN = NULL)

custom_annotation_wrapper(custom_function)

max_movement_detector(data, time_window_length = 10, ...)

median_movement_detector(data, time_window_length = 10, ...)

sum_movement_detector(data, time_window_length = 10, ...)
}
\arguments{
\item{func}{Aggregating function (max, min, median, mean, etc)}

\item{feature}{Name of a column in the sqlite3 file e.g. xy_dist_log10x1000}

\item{statistic}{Name of the column resulting from aggregation e.g. max_movement}

\item{score}{Name of the column providing a score i.e. category to the statistic e.g. micromovement
score is usually a binary variable i.e. TRUE/FALSE}

\item{preproc_FUN}{Optional, function to preprocess the input before computing the feature
(if the data needs some transformation like reverting xy_dist_log10x1000 back to a distance)}

\item{custom_function}{function used to produce the custom annotation}

\item{data}{\link{data.table} containing behavioural variable from or one multiple animals.
When it has a key, unique values, are assumed to represent unique individuals (e.g. in a \link{behavr} table).
Otherwise, it analysis the data as coming from a single animal. \code{data} must have a column \code{t} representing time.}

\item{time_window_length}{Size of non overlapping time bins, in seconds}

\item{...}{Extra arguments to be passed to \code{custom_function}.}
}
\value{
a \link{behavr} table similar to \code{data} with additional variables/annotations.
The resulting data will only have one data point every \code{time_window_length} seconds.
}
\description{
This function gives aggregates a variable of interest in a custom way
All datapoints in every time_window_length seconds is aggregated into a single datapoint
}
\details{
The default \code{time_window_length} is 300 seconds -- it is also known as the "5-minute rule".
}
