% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cycles_stats}
\alias{cycles_stats}
\title{Get cycles related stats. Number of cycles, total duration, average duration.}
\usage{
cycles_stats(e)
}
\arguments{
\item{e}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\value{
A named vector.
}
\description{
Get cycles related stats. Number of cycles, total duration, average duration.
}
