% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_annotation.R
\name{distance_sum_enclosed}
\alias{distance_sum_enclosed}
\title{A function to compute the distance traversed by an animal
Preprocess a raw ethoscope dataset by computing the sum of the number of pixels
traversed by an animal on each time bin}
\usage{
distance_sum_enclosed(d, time_window_length)
}
\value{
data.table of columns t and dist_sum
}
\description{
A function to compute the distance traversed by an animal
Preprocess a raw ethoscope dataset by computing the sum of the number of pixels
traversed by an animal on each time bin
}
