% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep-annotation.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Compute euclidean distance between two points}
\usage{
euclidean_distance(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
Compute euclidean distance between two points
}
