% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-functions.R
\name{generic_transition}
\alias{generic_transition}
\title{Compute a generatic transition probability fucntion}
\usage{
generic_transition(state0, state1)
}
\arguments{
\item{state0}{First state in the transtition}

\item{state1}{Second state in the transition}
}
\value{
A function that computes P(state0|state1) in a sleep trace
}
\description{
A closure that creates a probability transition function based on the
transition passed in args 1 and 2
}
\details{
The resulting function takes a trace of states where 1 encodes asleep and 0 awake
}
