% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_overlapping_duration}
\alias{get_overlapping_duration}
\title{Get the duration of x labels events overlapping y labels events.}
\usage{
get_overlapping_duration(x, y, e)
}
\arguments{
\item{x}{label}

\item{y}{stages}

\item{e}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\description{
Get the duration of x labels events overlapping y labels events.
}
