% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_overlapping_events}
\alias{get_overlapping_events}
\title{Filter events overlapping specified events.}
\usage{
get_overlapping_events(events, x, y)
}
\arguments{
\item{events}{Events.}

\item{x}{Events labels overlapping.}

\item{y}{Events labels to be overlapped.}
}
\value{
Dataframe of x events overlapping y events.
}
\description{
Filter events overlapping specified events.
}
