% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{hypnogram}
\alias{hypnogram}
\title{Filter an events dataframe to keep hypnogram.}
\usage{
hypnogram(events, labels = c("N3", "N2", "N1", "REM", "AWA"))
}
\arguments{
\item{events}{events dataframe.}

\item{labels}{events labels in data.}
}
\value{
hypnogram dataframe.
}
\description{
Filter an events dataframe to keep hypnogram.
}
