% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{hypnogram_band_powers}
\alias{hypnogram_band_powers}
\title{Compute bands powers from a full hypnogram.}
\usage{
hypnogram_band_powers(record, channel, bands = list(c(0.5, 3.5), c(3.5,
  8), c(8, 12), c(12, 30), c(30, 40), c(4, 40), c(0.5, 40)),
  bands_names = c("delta", "theta", "alpha", "beta", "gamma1",
  "denominator", "broadband"), normalize = c(4, 40))
}
\arguments{
\item{record}{record.}

\item{channel}{channel to split and hyp.}

\item{bands}{bands. Default "delta","theta","alpha","beta", "gamma1", "denominator", "broadband".}

\item{bands_names}{band_names.}

\item{normalize}{normalize.}
}
\value{
a df.
}
\description{
Compute bands powers from a full hypnogram.
}
