% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n1_duration}
\alias{n1_duration}
\title{Sums up N1 stages duration from an events dataframe to get total N1 duration in minutes.}
\usage{
n1_duration(events)
}
\arguments{
\item{events}{Events dataframe.}
}
\value{
Total duration of N1 sleep in minutes.
}
\description{
Sums up N1 stages duration from an events dataframe to get total N1 duration in minutes.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N1","N1")
n1_duration(events)
}
