% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n3_duration}
\alias{n3_duration}
\title{Sums up N3 stages duration from an events dataframe to get total N3 duration in minutes.}
\usage{
n3_duration(events)
}
\arguments{
\item{events}{Events dataframe.}
}
\value{
total duration of N3 sleep in minutes.
}
\description{
Sums up N3 stages duration from an events dataframe to get total N3 duration in minutes.
}
\examples{
events <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830),origin = "1970-01-01"))
events$end <- as.POSIXlt(c(1536967830,1536967860), origin = "1970-01-01")
events$event = c("N3","N3")
n3_duration(events)
}
