% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philrob.R
\name{philrob_default_parms}
\alias{philrob_default_parms}
\title{Default parameters of Phillips and Robinson's model}
\usage{
philrob_default_parms()
}
\value{
The default parameters for Strogatz's model
}
\description{
Loads the parameters used in Phillips and Robinson's model model
}
\details{
Phillips AJK, Robinson PA.
A Quantitative Model of Sleep-Wake Dynamics Based on the Physiology of the Brainstem Ascending Arousal System.
J Biol Rhythms. 2007 Apr 29;22(2):167–79. Available from: http://journals.sagepub.com/doi/10.1177/0748730406297512
}
\examples{
parms <- philrob_default_parms()
}
\references{
Phillips AJK, Robinson PA.
A Quantitative Model of Sleep-Wake Dynamics Based on the Physiology of the Brainstem Ascending Arousal System.
J Biol Rhythms \link{Internet}. 2007 Apr 29;22(2):167–79. Available from: http://journals.sagepub.com/doi/10.1177/0748730406297512
}
\seealso{
\code{\link{philrob}, \link{dphilrob}}
}
\author{
Pablo Rodríguez-Sánchez (\url{https://pabrod.github.io})
}
