% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{plot_hypnogram}
\alias{plot_hypnogram}
\title{Draw a histogram with ggplot2.}
\usage{
plot_hypnogram(events, labels = c("N3", "N2", "N1", "REM", "AWA"))
}
\arguments{
\item{events}{events dataframe.}

\item{labels}{hyp labels.}
}
\value{
a ggplot.
}
\description{
Draw a histogram with ggplot2.
}
