% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{plot_transitions}
\alias{plot_transitions}
\title{Get transitions graph}
\usage{
plot_transitions(hypnogram, height = "500px", width = "100\%")
}
\arguments{
\item{hypnogram}{hypnogram}

\item{height}{height}

\item{width}{width}
}
\value{
list.
}
\description{
Get transitions graph
}
