% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pos_stats}
\alias{pos_stats}
\title{Get position g events related stats in a named vector.}
\usage{
pos_stats(e, ss = c("N1", "N2", "N3", "REM"))
}
\arguments{
\item{e}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}

\item{ss}{sleep stages}
}
\value{
Total snoring count.
}
\description{
\code{snoring_stats} computes snoring related stats.
}
\examples{
e <- data.frame(begin = as.POSIXlt(c(1536967800,1536967830,1536967810),origin = "1970-01-01"))
e$end <- as.POSIXlt(c(1536967830,1536967860,1536967820), origin = "1970-01-01")
e$event = c("N3","N3","back")
pos_stats(e)
}
