% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_all_events_isruc}
\alias{read_all_events_isruc}
\title{Read all ISRUC sleep scorings (1)}
\usage{
read_all_events_isruc(target)
}
\arguments{
\item{target}{ISRUC target.}
}
\value{
A dataframe of scored events.
}
\description{
Read all ISRUC sleep scorings (1)
}
