% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_events_isruc}
\alias{read_events_isruc}
\title{Read ISRUC sleep scoring}
\usage{
read_events_isruc(dir, scoringNum)
}
\arguments{
\item{dir}{character, ISRUC record directory.}

\item{scoringNum}{Scoring number in database.}
}
\value{
A dataframe of scored stages.
}
\description{
Read ISRUC sleep scoring
}
