% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_events_sleepedfx}
\alias{read_events_sleepedfx}
\title{Read a SleepEDFX events file EDF+}
\usage{
read_events_sleepedfx(path, update = TRUE)
}
\arguments{
\item{path}{EDF+ path}

\item{update}{merge N3 and N4 or not}
}
\value{
A dataframe of scored events.
}
\description{
Read a SleepEDFX events file EDF+
}
