% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{read_isruc_metadata}
\alias{read_isruc_metadata}
\title{Read ISRUC-Sleep dataset records metadata.}
\usage{
read_isruc_metadata(dir)
}
\arguments{
\item{dir}{ISRUC dataset directory.}
}
\value{
A dataframe containing records metadata.
}
\description{
Read all the subgroups XLSX metadata files and concatenates the dataframes into one.
}
