% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rem_stats}
\alias{rem_stats}
\title{Get REM related stats.}
\usage{
rem_stats(e)
}
\arguments{
\item{e}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}
}
\value{
A named vector.
}
\description{
Get REM related stats.
}
