% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep-annotation.R
\name{sleep_annotation}
\alias{sleep_annotation}
\alias{sleep_dam_annotation}
\title{Infer sleep from immobility bouts}
\usage{
sleep_annotation(data, time_window_length = 10, min_time_immobile = 300,
  motion_detector_FUN = max_velocity_detector, ...)

sleep_dam_annotation(data, time_window_length = 60, min_time_immobile = 300)
}
\arguments{
\item{data}{\link{behavr} table containing behavioural variable from multiple (or one) animals.
It must contain, at least, the column \code{t} in addition to the variables needed for motion classification.}

\item{time_window_length}{number of seconds to be used by the motion classifier.
This corresponds to the sampling period of the output data.}

\item{min_time_immobile}{Minimal duration (in s) of a sleep bout.
Immobility bouts longer or equal to this value are considered as sleep.}

\item{motion_detector_FUN}{function used to classify movement}

\item{...}{extra arguments to be passed to \code{motion_classifier_FUN}.}
}
\value{
a \link{behavr} table similar to \code{data} with additional variables/annotations (i.e. \code{moving} and \code{asleep}).
The resulting data will only have one data point every \code{time_window_length} seconds.
}
\description{
This function first uses a motion classifier to decide whether an animal is moving during a given time window.
Then, it defines sleep as contiguous immobility for a minimal duration.
}
\details{
The default \code{time_window_length} is 300 seconds also known as the "5min rule".
\code{sleepAnnotation} is typically used for ethoscope data, whilst \code{sleepDAMAnnotation} only works on DAM data.
These function is \emph{rarely used directly}, but rather passed as an argument to a data loading function,
so that analysis can be performed on the go.
}
\examples{
#todo
}
