% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{split_signal}
\alias{split_signal}
\title{Split signal following hypnogram}
\usage{
split_signal(signal, hypnogram, sRate)
}
\arguments{
\item{signal}{signal.}

\item{hypnogram}{hypnogram df.}

\item{sRate}{sRate sig.}
}
\value{
list.
}
\description{
Split signal following hypnogram
}
