% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tm}
\alias{tm}
\title{Computes sleep transition matrix from selected stages.}
\usage{
tm(e, l = c("AWA", "N1", "N2", "N3", "REM"))
}
\arguments{
\item{e}{Events dataframe. Dataframe must have \code{begin} (\code{POSIXt}), \code{end} (\code{POSIXt}) and \code{event} (\code{character}) columns.}

\item{l}{Stages labels. Defaults to \code{c("AWA", "N1", "N2", "N3", "REM")}}
}
\value{
A matrix.
}
\description{
Sleep transition matrix is the matrix containing probabilities of next stage transition for a given stage.
}
