% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tts}
\alias{tts}
\title{Sums up REM, N1, N2 and N3 stages duration from an events dataframe to get Time To Sleep duration in minutes.}
\usage{
tts(events)
}
\arguments{
\item{events}{Events dataframe. Must contain `begin`, `end` and `events`.}
}
\value{
Time To Sleep in minutes.
}
\description{
Sums up REM, N1, N2 and N3 stages duration from an events dataframe to get Time To Sleep duration in minutes.
}
